resolved more of Vaughn's comments
