+++ vpr/src/route/route_timing.cpp
@@ -50,6 +50,11 @@ enum class RouterCongestionMode {
     CONFLICTED
 };
 
+typedef enum router_breakpoint_type {
+    BP_ROUTE_ITER,
+    BP_NET_ID
+} bp_router_type;
+
 struct RoutingMetrics {
     size_t used_wirelength = 0;
 
@@ -164,7 +169,7 @@ static void init_net_delay_from_lookahead(const RouterLookahead& router_lookahea
                                           ClbNetPinsMatrix<float>& net_delay);
 
 //debug related functions
-void update_router_info_and_check(char type, int net_id);
+void update_router_info_and_check_bp(bp_router_type type, int net_id);
 void send_current_info_r();
 
 // The reason that try_timing_driven_route_tmpl (and descendents) are being
@@ -423,7 +428,7 @@ bool try_timing_driven_route_tmpl(const t_router_opts& router_opts,
 
             if (was_rerouted) {
                 rerouted_nets.push_back(net_id);
-                update_router_info_and_check('n', size_t(net_id));
+                update_router_info_and_check_bp(BP_NET_ID, size_t(net_id));
             }
         }
 
@@ -549,7 +554,7 @@ bool try_timing_driven_route_tmpl(const t_router_opts& router_opts,
         if (legal_convergence_count >= router_opts.max_convergence_count
             || router_iteration_stats.connections_routed == 0
             || early_reconvergence_exit_heuristic(router_opts, itry_since_last_convergence, timing_info, best_routing_metrics)) {
-            update_router_info_and_check('r', -1);
+            update_router_info_and_check_bp(BP_ROUTE_ITER, -1);
             break; //Done routing
         }
 
@@ -558,7 +563,7 @@ bool try_timing_driven_route_tmpl(const t_router_opts& router_opts,
          */
         if (itry == 1 && early_exit_heuristic(router_opts, wirelength_info)) {
             //Abort
-            update_router_info_and_check('r', -1);
+            update_router_info_and_check_bp(BP_ROUTE_ITER, -1);
             break;
         }
 
@@ -568,14 +573,14 @@ bool try_timing_driven_route_tmpl(const t_router_opts& router_opts,
 
             if (!std::isnan(est_success_iteration) && est_success_iteration > abort_iteration_threshold) {
                 VTR_LOG("Routing aborted, the predicted iteration for a successful route (%.1f) is too high.\n", est_success_iteration);
-                update_router_info_and_check('r', -1);
+                update_router_info_and_check_bp(BP_ROUTE_ITER, -1);
                 break; //Abort
             }
         }
 
         if (itry == 1 && router_opts.exit_after_first_routing_iteration) {
             VTR_LOG("Exiting after first routing iteration as requested\n");
-            update_router_info_and_check('r', -1);
+            update_router_info_and_check_bp(BP_ROUTE_ITER, -1);
             break;
         }
 
@@ -1996,17 +2001,19 @@ static void init_net_delay_from_lookahead(const RouterLookahead& router_lookahea
     }
 }
 
-//updates router iteration information and checks for breakpoints
-//stops if a breakpoint is encountered
-void update_router_info_and_check(char type, int net_id) {
+//updates router iteration information and checks for router iteration and net id breakpoints
+//stops after the specified router iteration or net id is encountered
+void update_router_info_and_check_bp(bp_router_type type, int net_id) {
     t_draw_state* draw_state = get_draw_state_vars();
     if (draw_state->list_of_breakpoints.size() != 0) {
-        if (type == 'r')
+        if (type == BP_ROUTE_ITER)
             current_info_r.router_iter++;
-        else if (type == 'n')
+        else if (type == BP_NET_ID)
             current_info_r.net_id = net_id;
         send_current_info_r();
-        f_router_debug = check_for_breakpoints();
+
+        //checks for breakpoints and prints appropriate message if breakpoint was encountered
+        f_router_debug = check_for_breakpoints(false);
         if (f_router_debug) {
             breakpoint_info_window(get_current_info_b().bp_description, current_info_r);
             update_screen(ScreenUpdatePriority::MAJOR, "Breakpoint Encountered", ROUTING, nullptr);
